requirements.txt
arxiv
fastapi
uvicorn
pytz
apscheduler
python-dotenv
google-generativeai
httpx